requirements.txt
fastapi==0.109.0
uvicorn==0.22.0
psycopg2==2.9.9
google-cloud-translate==3.21.1
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
openpyxl==3.1.5
dotenv==0.9.9